Desktop_brixxbee/requirements.txt
SpeechRecognition
pyttsx3
PyAudio
requests
openai
customtkinter
pillow
